requirements.txt
#放的是所需要用到的工具版本
pytest
request